refactor: move REPL logic to separate module
